Created UserProfilePictureRepository and moved the functionality in there, made service check for errors
