+++ Application/Services/UserService.cs
@@ -1,4 +1,5 @@
-﻿using Core.Entities;
+﻿using Azure;
+using Core.Entities;
 using Core.Exceptions;
 using Core.Interfaces;
 using Core.Interfaces.Repositories;
@@ -56,7 +57,6 @@ namespace Application.Services
         {
             _userRepository.Delete(user);
             await _userRepository.UpdateAsync(user);
-
             await _userProfilePictureService.DeleteAsync(user.ProfilePicture!);
 
             _loggerManager.LogInfo($"Successfully deleted the user with id {user.Id}");
