+++ DataAccess/Repositories/Configuration/ApplicationRepositoriesConfiguration.cs
@@ -21,6 +21,7 @@ namespace DataAccess.Repositories.Configuration
             services.AddScoped<IUserSpecializationRepository, UserSpecializationRepository>();
             services.AddScoped<IArticleRepository, ArticleRepository>();
             services.AddScoped<ICommentRepository, CommentRepository>();
+            services.AddScoped<IUserProfilePictureRepository, UserProfilePictureRepository>();
         }
     }
 }
